Modify server file structure
Include seperate files for controllers and models. Routers are retained in index.js. This file structure makes the server easier to maintain.
