+++ server/models/users.js
@@ -1,0 +1,16 @@
+const userList = new Map([
+  ['admin', {
+    id: 'Integer',
+    firstname: 'String',
+    lastname: 'String',
+    othernames: 'String',
+    email: 'String',
+    phoneNumber: 'String',
+    username: 'admin',
+    registered: 'Date',
+    isAdmin: true,
+    password: 'encoded password',
+  }],
+]);
+
+module.exports = userList;
